# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Onboarding Platform API
  version: 1.0.0
  description: Backend API for AI Onboarding & Training platform
servers:
  - url: /api/v1
    description: V1 base path (documentation only)
...
paths:
...
  /api/v1/roadmaps:
    post:
      summary: Create roadmap
      description: Creates a new learning roadmap.
      tags: [ roadmaps ]
      operationId: createRoadmap
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        description: Request payload for the createRoadmap.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRoadmapRequestV1' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/RoadmapResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
  /api/v1/roadmaps/search:
    post:
      summary: Search roadmaps
      description: Returns a page of roadmaps visible to the authenticated user.
      tags: [ roadmaps ]
      operationId: searchRoadmaps
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        description: Roadmap search, filter, sort, and pagination criteria.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SearchRoadmapsV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RoadmapsListResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
  /api/v1/roadmaps/count:
    post:
      summary: Count roadmaps
      description: Returns the total number of roadmaps matching the given filter without fetching or paginating them — for cheap tab-count display while another tab's list is active. page, size, sort, and direction are ignored.
      tags: [ roadmaps ]
      operationId: countRoadmaps
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        description: Same filter fields as searchRoadmaps; page, size, sort, and direction are ignored.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SearchRoadmapsV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CountResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
  /api/v1/roadmaps/{id}:
    put:
      summary: Update roadmap
      description: Updates roadmap title and description.
      tags: [ roadmaps ]
      operationId: updateRoadmap
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the updateRoadmap.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRoadmapRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RoadmapResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
    delete:
      summary: Delete roadmap
      description: Deletes a roadmap by id.
      tags: [ roadmaps ]
      operationId: deleteRoadmap
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkIdResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/roadmaps/{id}/assignments:
    post:
      summary: Assign roadmap
      description: Assigns a roadmap to one or more users.
      tags: [ roadmaps ]
      operationId: assignRoadmap
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the assignRoadmap.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignmentRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AssignmentResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/roadmaps/{id}/assignees:
    get:
      summary: List roadmap assignees
      description: Lists learners currently assigned to (enrolled in) a roadmap, for assignment management.
      tags: [ roadmaps ]
      operationId: listRoadmapAssignees
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/LearningAssigneesResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/roadmaps/{id}/assignments/revoke:
    post:
      summary: Revoke roadmap assignments
      description: Revokes one or more learners' roadmap assignment (enrollment) in one request.
      tags: [ roadmaps ]
      operationId: revokeRoadmapAssignments
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the revokeRoadmapAssignments.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignmentRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/roadmaps/{id}/enrollment:
    post:
      summary: Enroll in roadmap
      description: Enrolls the current user in a roadmap.
      tags: [ roadmaps ]
      operationId: enrollRoadmap
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/EnrollmentResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
    delete:
      summary: Unenroll from roadmap
      description: Removes the current user's enrollment from a roadmap.
      tags: [ roadmaps ]
      operationId: unenrollRoadmap
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/roadmaps/{id}/team-assignments:
    get:
      summary: List roadmap team assignments
      description: Lists the teams a roadmap is standing-assigned to, visible to the caller.
      tags: [ roadmaps ]
      operationId: listRoadmapTeamAssignments
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RoadmapTeamAssignmentsListResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
    post:
      summary: Assign roadmap to team
      description: Assigns a roadmap to a team, identified by its lead's user id, and enrolls every current team member.
      tags: [ roadmaps ]
      operationId: assignRoadmapToTeam
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the assignRoadmapToTeam.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignRoadmapToTeamRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RoadmapTeamAssignmentResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/roadmaps/{id}/team-assignments/{leadUserId}:
    delete:
      summary: Unassign roadmap from team
      description: Removes a roadmap's standing assignment to a team. Existing member enrollments are left untouched.
      tags: [ roadmaps ]
      operationId: unassignRoadmapFromTeam
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: leadUserId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/roadmaps/{id}/group-assignments:
    get:
      summary: List roadmap group assignments
      description: Lists the groups a roadmap is standing-assigned to, restricted to groups the caller can manage.
      tags: [ roadmaps ]
      operationId: listRoadmapGroupAssignments
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RoadmapGroupAssignmentsListResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
    post:
      summary: Assign roadmap to group
      description: Creates or updates the standing assignment of a roadmap to a group, replacing any existing grade filter, and enrolls every currently matching group member.
      tags: [ roadmaps ]
      operationId: assignRoadmapToGroup
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the assignRoadmapToGroup.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignRoadmapToGroupRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RoadmapGroupAssignmentResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/roadmaps/{id}/group-assignments/preview:
    post:
      summary: Preview roadmap group assignment
      description: Previews how many current group members a grade filter would match, without creating an assignment or enrolling anyone.
      tags: [ roadmaps ]
      operationId: previewRoadmapGroupAssignment
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the previewRoadmapGroupAssignment.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoadmapGroupAssignmentPreviewRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/RoadmapGroupAssignmentPreviewResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
  /api/v1/roadmaps/{id}/group-assignments/{groupId}:
    delete:
      summary: Unassign roadmap from group
      description: Removes a roadmap's standing assignment to a group. Existing member enrollments and progress are left untouched.
      tags: [ roadmaps ]
      operationId: unassignRoadmapFromGroup
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: groupId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OkResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedV1:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
    ForbiddenV1:
      description: Authenticated but lacking permission
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
    NotFoundV1:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
...
  schemas:
...
    CountResponseV1:
      description: Bounded count-only payload for a search filter, without fetching or paginating the matching rows — used for tab-count display independent of which tab's full list is active.
      type: object
      additionalProperties: false
      required: [ totalElements ]
      properties:
        totalElements:
          description: Total number of items matching the filter.
          type: integer
          format: int64
    SortDirectionV1:
      description: Sort direction applied to a list query's sort field.
      type: string
      enum: [ asc, desc ]
...
    RoadmapSortFieldV1:
      description: Sortable field for roadmap list queries.
      type: string
      enum: [ createdAt, updatedAt, title ]
...
    SearchRoadmapsV1:
      description: Roadmap search, filter, sort, and pagination criteria.
      type: object
      additionalProperties: false
      properties:
        page:
          description: Zero-based page number to return.
          type: integer
          format: int32
          minimum: 0
          default: 0
        size:
          description: Maximum number of roadmaps to return per page.
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          default: 20
        sort:
          description: Field used to sort the roadmap list.
          allOf:
            - $ref: '#/components/schemas/RoadmapSortFieldV1'
          default: createdAt
        direction:
          description: Sort direction applied to the sort field.
          allOf:
            - $ref: '#/components/schemas/SortDirectionV1'
          default: desc
        query:
          description: Free-text search matched against roadmap title, description, creator name, and tags.
          type: string
        tags:
          description: Restricts results to roadmaps tagged with every one of the given values.
          type: array
          items: { type: string }
        createdByUserId:
          description: Restricts results to roadmaps created by this internal user id.
          type: integer
          format: int64
        assignedToMe:
          description: When true, restricts results to roadmaps the current viewer is enrolled in.
          type: boolean
...
    ApiErrorV1:
      description: Standard API error payload.
      type: object
      additionalProperties: false
      required: [ code, message, timestamp ]
      properties:
        code:
          description: Machine-readable code value.
          type: string
        message:
          description: Human-readable message text.
          type: string
        timestamp:
          description: Timestamp for the event or error.
          type: string
          format: date-time
        correlationId:
          description: Correlation identifier for tracing.
          type: string
        parameters:
          description: Parameters for Api Error.
          type: array
          items: { $ref: '#/components/schemas/ValidationParameterV1' }
...
    OkResponseV1:
      description: Ok Response payload.
      type: object
      additionalProperties: false
      required: [ ok ]
      properties:
        ok:
          description: Indicates whether the operation succeeded.
          type: boolean
    OkIdResponseV1:
      description: Ok Id Response payload.
      type: object
      additionalProperties: false
      required: [ ok, id ]
      properties:
        ok:
          description: Indicates whether the operation succeeded.
          type: boolean
        id:
          description: Unique identifier.
          type: integer
          format: int64
...
    AssignmentRequestV1:
      description: Assignment Request payload.
      type: object
      additionalProperties: false
      required: [ userIds ]
      properties:
        userIds:
          description: User Ids for Assignment Request.
          type: array
          items: { type: integer, format: int64 }
...
    LearningAssigneesResponseV1:
      description: Learning Assignees Response payload.
      type: object
      additionalProperties: false
      required: [ assignees ]
      properties:
        assignees:
          description: Learners currently assigned to the lesson or roadmap.
          type: array
          items: { $ref: '#/components/schemas/LearningAssigneeV1' }
...
    AssignRoadmapToTeamRequestV1:
      description: Assign Roadmap To Team Request payload.
      type: object
      additionalProperties: false
      required: [ leadUserId ]
      properties:
        leadUserId:
          description: Team lead user id identifying the target team.
          type: integer
          format: int64
...
    RoadmapTeamAssignmentResponseV1:
      description: Roadmap Team Assignment Response payload.
      type: object
      additionalProperties: false
      required: [ ok, assignment, enrollments ]
      properties:
        ok:
          description: Indicates whether the operation succeeded.
          type: boolean
        assignment: { $ref: '#/components/schemas/RoadmapTeamAssignmentV1' }
        enrollments:
          description: Enrollment records created for current team members.
          type: array
          items: { $ref: '#/components/schemas/RoadmapAssignmentEnrollmentV1' }
    RoadmapTeamAssignmentsListResponseV1:
      description: Roadmap Team Assignments List Response payload.
      type: object
      additionalProperties: false
      required: [ assignments ]
      properties:
        assignments:
          description: Assignments for Roadmap Team Assignments List Response.
          type: array
          items: { $ref: '#/components/schemas/RoadmapTeamAssignmentV1' }
...
    AssignRoadmapToGroupRequestV1:
      description: Assign Roadmap To Group Request payload.
      type: object
      additionalProperties: false
      required: [ groupId ]
      properties:
        groupId:
          description: Target group id.
          type: integer
          format: int64
        gradeIds:
          description: Grade ids to narrow the assignment to. Omitted or empty means every member.
          type: array
          items: { type: integer, format: int64 }
    RoadmapGroupAssignmentResponseV1:
      description: Roadmap Group Assignment Response payload.
      type: object
      additionalProperties: false
      required: [ ok, assignment, enrollments ]
      properties:
        ok:
          description: Indicates whether the operation succeeded.
          type: boolean
        assignment: { $ref: '#/components/schemas/RoadmapGroupAssignmentV1' }
        enrollments:
          description: Enrollment records created or refreshed for currently matching group members.
          type: array
          items: { $ref: '#/components/schemas/RoadmapAssignmentEnrollmentV1' }
    RoadmapGroupAssignmentsListResponseV1:
      description: Roadmap Group Assignments List Response payload.
      type: object
      additionalProperties: false
      required: [ assignments ]
      properties:
        assignments:
          description: Assignments for Roadmap Group Assignments List Response.
          type: array
          items: { $ref: '#/components/schemas/RoadmapGroupAssignmentV1' }
    RoadmapGroupAssignmentPreviewRequestV1:
      description: Roadmap Group Assignment Preview Request payload.
      type: object
      additionalProperties: false
      required: [ groupId ]
      properties:
        groupId:
          description: Group id to preview against.
          type: integer
          format: int64
        gradeIds:
          description: Candidate grade ids. Omitted or empty means every member.
          type: array
          items: { type: integer, format: int64 }
...
    RoadmapGroupAssignmentPreviewResponseV1:
      description: Roadmap Group Assignment Preview Response payload.
      type: object
      additionalProperties: false
      required: [ preview ]
      properties:
        preview:
          description: The computed preview.
          allOf:
            - $ref: '#/components/schemas/RoadmapGroupAssignmentPreviewV1'
    AssignmentResponseV1:
      description: Assignment Response payload.
      type: object
      additionalProperties: false
      required: [ ok, enrollments ]
      properties:
        ok:
          description: Indicates whether the operation succeeded.
          type: boolean
        enrollments:
          description: Enrollments for Assignment Response.
          type: array
          items:
            oneOf:
              - { $ref: '#/components/schemas/AssignmentEnrollmentV1' }
              - { $ref: '#/components/schemas/RoadmapAssignmentEnrollmentV1' }
    EnrollmentResponseV1:
      description: Enrollment Response payload.
      type: object
      additionalProperties: false
      required: [ ok, enrollment ]
      properties:
        ok:
          description: Indicates whether the operation succeeded.
          type: boolean
        enrollment:
          description: Enrollment for Enrollment Response.
          oneOf:
            - { $ref: '#/components/schemas/LessonEnrollmentV1' }
            - { $ref: '#/components/schemas/RoadmapEnrollmentV1' }
        completedRoadmaps:
          description: Completed Roadmaps for Enrollment Response.
          type: array
          items: { $ref: '#/components/schemas/CompletedRoadmapSummaryV1' }
...
    RoadmapsListResponseV1:
      description: Roadmaps List Response payload.
      type: object
      additionalProperties: false
      required: [ roadmaps, page ]
      properties:
        roadmaps:
          description: Roadmaps for Roadmaps List Response.
          type: array
          items: { $ref: '#/components/schemas/RoadmapV1' }
        page:
          description: Pagination metadata for this result set.
          allOf:
            - $ref: '#/components/schemas/PageInfoV1'
    CreateRoadmapRequestV1:
      type: object
      additionalProperties: false
      required: [ title, lessonIds ]
      properties:
        title:
          description: Title text.
          type: string
        description:
          type: string
          title: Human-readable description text.
        lessonIds:
          description: Ordered lesson identifiers.
          type: array
          items: { type: integer, format: int64 }
        tags:
          description: Tags for the roadmap.
          type: array
          items: { type: string }
    RoadmapResponseV1:
      description: Roadmap Response payload.
      type: object
      additionalProperties: false
      required: [ roadmap ]
      properties:
        roadmap:
          description: Primary roadmap label for the member.
          allOf:
            - $ref: '#/components/schemas/RoadmapV1'
    UpdateRoadmapRequestV1:
      type: object
      additionalProperties: false
      properties:
        title:
          description: Title text.
          type: string
        description:
          type: string
          title: Human-readable description text.
        lessonIds:
          description: Ordered lesson identifiers.
          type: array
          items: { type: integer, format: int64 }
        tags:
          description: Tags for the roadmap.
          type: array
          items: { type: string }
...